+++ src/extension.ts
@@ -1,12 +1,10 @@
 import * as vscode from 'vscode';
-import { browseRow, freezeCmd, freezeManual, installCmd, installManual, installRow } from './commands';
+import { browseRow, freezeCmd, installCmd, installRow } from './commands';
 
 export function activate(context: vscode.ExtensionContext) {
 	context.subscriptions.push(
 		installCmd,
 		freezeCmd,
-		installManual,
-		freezeManual,
 		installRow,
 		browseRow,
 	);