+++ src/templates.ts
@@ -7,36 +7,36 @@ function runPipCommand(path: string, command: string) {
     terminal.show();
 }
 
-export function focusTemplate(event: any, command: string, commandName: string) {
+export function focusTemplate(event: any, command: string) {
+    const runThat = (path: string) => runPipCommand(path, command);
     if (event) {
-        runPipCommand(event.fsPath, command);
+        runThat(event.fsPath);
     } else {
         const editor = vscode.window.activeTextEditor;
         if (editor !== undefined) {
-            runPipCommand(editor.document.uri.fsPath, command);
+            runThat(editor.document.uri.fsPath);
         } else {
-            vscode.commands.executeCommand(`${extName}.${commandName}-manual`);
+            runCommandManually(command);
         }
     }
 }
 
-export function manualTemplate(command: string) {
+async function runCommandManually(command: string) {
     const ignoreFocusOutSetting: boolean = vscode.workspace.getConfiguration(
         'pip-requirements'
-    ).get('ignoreFocusOut')!;
-    vscode.window.showInputBox({
+    ).get('ignoreFocusOut') || false;
+    let pathResponse = await vscode.window.showInputBox({
         title: 'Path to the requirements.txt file',
         value: defaultPath,
         placeHolder: defaultPath,
         ignoreFocusOut: ignoreFocusOutSetting
-    }).then(path => {
-        if (path !== undefined) {
-            if (path === '') {
-                path = defaultPath;
-            }
-            runPipCommand(path, command);
-        }
     });
+    if (pathResponse !== undefined) {
+        if (pathResponse === '') {
+            pathResponse = defaultPath;
+        }
+        runPipCommand(pathResponse, command);
+    }
 }
 
 // reexport templates with web support